NewQuote Functionality
New quote functionality. If you hit it, it will check if you have a
currentorder, if you do, clears out the cart and your local cache of the
current order then it will clear the filters on search and go there. If
no current order, go to search. Checks if it  is in xp.Status = DRAFT
and status = Unsubmitted if it is opens the modal dialog box.