Carriage Addition to XP
Login- now add s a cookie of the buyer's catalog which will contain
default carriage price.
Myquote - additions of modal controller for carriage missing. added
functionality to the myquote isvalid for review function which will
confirm the order has the xp required. also a function to check if the
carriage is set to go to review.
Delievery tpl is thje carriage additions.
Myquote html- shows a dummy reivew and submit quote for order if the
user hasnt chosen a carriage choice.